Move file routes out of index.ts, some cleanup
